photo descriptions generation separated on modules logic
